Blackbox macros: refresh type parameter names
This is a solution to the issue of type parameter shadowing.
Scalac issues a warning when we define a type parameter `A` if
there is already another one in scope with the same name.
This could be a problem for users with `-Xfatal-warnings` enabled.

The solution is to temporarily set fresh names for all type parameters,
type-check the tree and restore the original names afterwards.
It's a bit hacky, but self-contained.

Reverts type parameters which were renamed to work around the issue.