Fix DeriveMacros when a higher-kinded refined type is passed as a ref
One can pass a refined type `Hk[A] { type T }` as a reference:

```scala
Derive.functorK[Hk { type T = Int }]
```

and this currently breaks the derivation.

To resolve the problem we `etaExpand` the type constructor to
obtain a `PolyType` which is easier to work with.

As a cleanup, extract the initial normalization performed on
the algebra type parameter to a helper method.