Replace scalameta macro annotations with scala-reflect macros. (#45)
* Replace scalameta macro annotations with scala-reflect macros.

Migrate all the existing macro annotations to scala-reflect macros. Handle all the possible derivations (with/without abstract type members + summoning methods). See https://github.com/typelevel/cats-tagless/issues/37 for reference

* Fix DeriveMacros when a higher-kinded refined type is passed as a ref

One can pass a refined type `Hk[A] { type T }` as a reference:

```scala
Derive.functorK[Hk { type T = Int }]
```

and this currently breaks the derivation.

To resolve the problem we `etaExpand` the type constructor to
obtain a `PolyType` which is easier to work with.

As a cleanup, extract the initial normalization performed on
the algebra type parameter to a helper method.

* Remove legacy scalameta macro project

* Fix incompatibility with 2.13 (no AssignOrNamedArgs)

* Rename `monad` to `flatMap`

* Preserve type-bounds, kinds etc. of fully refined type params

* Add TODO explaining deficiencies in abstract type members computation

* Remove unused