Add Instrument.With that lets us instrument an algebra with a typeclass
The typeclass should have an instance for all (non-implicit) parameters
and also for the return type of every method.

It would be more flexible to allow two typeclasses for arguments and
return type but in practice this will be always `Show`, `Encoder` and
similar. On the other hand it would make the API too ugly.

Also add a `Trivial` typeclass that is always available.