Add Two New Error Related Methods To FunctorK.
This commit adds two new methods on the `FunctorK` typeclass, `transformError` and `flatTapOnError`.

For any `FunctorK[A]` for which the current `F` on `A[F]` is an instance of `MonadError[F, E]` (for some `E`) we can provide a method which allows for transformation on all `E` values for all methods in the given `A[F]`. This uses the identity `FunctionK` for invocations which do not yield an error and the `handleErrorWith` methods from `MonadError` for the error paths.

The primary motivating use case for these methods is creating a succinct way to define logging in the case of an error for all methods of an F-Algebra. For example with these changes it becomes possible to write something like this,

```scala
trait Logger[F[_]] {
    def error(message: String): F[Unit]
    def warn(message: String): F[Unit]
    def info(message: String): F[Unit]
}

@autoFunctorK
trait FAlgebra[F[_]] {
    def lookupUser(id: Id): F[User]
    def lookupOrder(orderId: OrderId): F[Order]
    def createOrder(order: Order): F[OrderId]
}

object FAlgebra {
    def simpleInstance[F[_]]: FAlgebra[F] = new FAlgebra[F]{
        override def lookupUser(id: Id): F[User] = ???
        override def lookupOrder(orderId: OrderId): F[Order] = ???
        override def createOrder(order: Order): F[OrderId] = ???
    }

    def withErrorLogging[F[_]: Sync](baseInstance: FAlgebra[F], logger: Logger[F]): FAlgebra[F] =
        baseInstance.flatTapOnError{
            case t: Throwable =>
                logger.error(t.getMessage)
        }
}

```